backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
sqlalchemy>=2.0,<3.0
python-dotenv>=1.0,<2.0
pydantic>=2.7,<3.0
pydantic-settings>=2.2,<3.0
psycopg2-binary>=2.9,<3.0
python-jose[cryptography]>=3.3,<4.0
passlib==1.7.4
bcrypt==4.3.0
email-validator>=2.1,<3.0
python-multipart>=0.0.9,<1.0
